Fix development language for Localizable.strings
Xcode wants development language come first in the localization
group, in fact it was picking Arabic despite the development
language set at the project level (a7ec2c6).

At the same time, it doesn't provide ANY WAY to fix it in the
UI. So I had to edit .pbxproj manually.