Drop redundant migration to automatic socket type
An old tunnelConfiguration without endpointProtocols will be
parsed to nil and therefore fall back to defaults, that is
AppConfiguration.shared.piaDefaultConfigurationBuilder. This will
override any former user UDP/TCP preference, but it's not a big
issue.

Such builder has endpointProtocols properly set to automatic, so:

- tunnelConfiguration will never be nil, it'll be set to defaults.
- Default endpointProtocols will never be empty.

As a consequence, the code will never be executed.